% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AlleleShare.R
\name{AlleleShare}
\alias{AlleleShare}
\title{Calculate numbers of sharing alleles each pair at each locus}
\usage{
AlleleShare(g,sep,replacement=FALSE
}
\arguments{
\item{g}{a dataframe of genotype data with rownames of sample ID and column names of markers.}

\item{sep}{allele separator in the imported genotype data. Note: when using the special character like "|", remember to protect it as "\\|"}

\item{replacement}{a logical variable. if it is TRUE, the pairs are sampled with replacement; if FALSE, the pairs are sampled without replacement.}
}
\value{
a matrix of numbers of shared alleles. Each row denotes each pair; Each column denotes each locus.
}
\description{
Calculate numbers of sharing alleles each pair at each locus
}
\details{
This function calculates the numbers of shared alleles between each pair of individuals for a dataset.
}
\examples{
df <- data.frame(SNP1=c("A/A","T/T","A/T","A/T","T/A","A/T","A/A","T/A","T/T","A/T"),
                STR1=c("12/12","13/14","13/13","14/15","15/13","13/14","14/13","12/12","14/14","15/15"))
AlleleShare_Table(g,"/",replacement=FALSE)

}
