% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counta.R
\name{counta}
\alias{counta}
\title{Simple count including zero###}
\usage{
counta(z, y)
}
\arguments{
\item{z}{a vector you would like to check}

\item{y}{an element you would like to count.(Even it is not included in z)}
}
\value{
the times that y appears in z
}
\description{
Simple count including zero###
}
\details{
This function counts how many the assigned elements there are in one vector.
}
\examples{
z <-rbinom(20,1,0.5)
counta(z,0)

}
