% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DistHetero.R
\name{DistHetero}
\alias{DistHetero}
\title{Build Expected Distribution of Numbers of Heterozygous Loci}
\usage{
DistHetero(H)
}
\arguments{
\item{H}{a vector of average heterozygosity of each locus}
}
\value{
a dataframe of expected density on each possible total number of heterozygous loci.
}
\description{
Build Expected Distribution of Numbers of Heterozygous Loci
}
\details{
This function build the expected distribution of numbers of heterozygous loci for known heterozygosity of each loci.
}
\examples{
DistHetero(runif(10))

}
\references{
Chakraborty, R. (1981, ISSN:0016-6731)
}
