% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FreqAlleleShare.R
\name{FreqAlleleShare}
\alias{FreqAlleleShare}
\title{Build Observed Distribution of No. of Shared Alleles}
\usage{
FreqAlleleShare(AS)
}
\arguments{
\item{AS}{a matrix of number of shared alleles, made up with 0, 1 and 2, outcome of function "AlleleShare_Table". Rows for individuals, and columns for markers.}
}
\value{
a dataframe of frequencies of each number of shared alleles(from 0 to 2*N0. of loci)
}
\description{
Build Observed Distribution of No. of Shared Alleles
}
\details{
This function build the observed distributions from observed Allele Share table, made up of 0,1 and 2.
}
\examples{
AS<-matrix(sample(c(0:2),20,replace=TRUE,prob=c(0.3,0.3,0.4)),nrow=5)
FreqAlleleShare(AS)
}
