% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixexample-data.R
\docType{data}
\name{mixexample}
\alias{mixexample}
\title{Genotype Data from A Selected Mix Panel}
\format{
A dataframe with 2504 observations on 100 variables. This dataframe is the phased genotype files for 100 variants (including SNPs and STRs) for 2504 individuals.
}
\source{
\href{https://github.com/ice4prince/mixIndependR/tree/main/data}
}
\usage{
data(mixexample)
}
\description{
This dataset is the phased genpotypes for a mix panel with 100 variants.
These variants are selected from the reference haplotype data of Gymrek's lab (see Reference).
This is a sample with 2504 individuals.
}
\examples{
data(mixexample)
}
\references{
Saini et al. (2018). Nat Commun 9(1): 4397.
(\href{https://gymreklab.com/2018/03/05/snpstr_imputation.html})
}
\keyword{datasets}
