% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vcf_gt.R
\name{read_vcf_gt}
\alias{read_vcf_gt}
\title{Import genotype data from vcf files/}
\usage{
read_vcf_gt(x)
}
\arguments{
\item{x}{The vcf file with its directory}
}
\value{
a list contains the genotype and allele status.
}
\description{
Import genotype data from vcf files/
}
\details{
This function extract the genotypes and allele status from a vcf file.
}
\examples{
\dontrun{
df<-read_vcf_gt("~/x.vcf")
}
}
