% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulate_DistK.R
\name{Simulate_DistK}
\alias{Simulate_DistK}
\title{Generate a Bundle of Simulated distributions for No. of heterozygous loci with known heterozygosites}
\usage{
Simulate_DistK(H,m,t)
}
\arguments{
\item{H}{a vector of average heterozygosity of each loci. Length of H is the number of loci.}

\item{m}{the sample size you want, usually similar to the real sample size.}

\item{t}{the number of samples you want to build}
}
\value{
a matrix of frequencies of No. of Heterozygous Loci. Each row denotes each simulated sample; Each column denotes each No. of Heterozygous loci, from 0 to length of H.
}
\description{
Generate a Bundle of Simulated distributions for No. of heterozygous loci with known heterozygosites
}
\details{
This function generates multinomial distribution for loci known the heterozygosity and build the simulated distribution for no. of heterozygous loci.
}
\examples{
Simulate_DistK(runif(10),500,100)

}
