% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Heterozygous.R
\name{Heterozygous}
\alias{Heterozygous}
\title{Test heterozygosity at each locus}
\usage{
Heterozygous(x,sep)
}
\arguments{
\item{x}{a dataset of genotypes with rownames of sample ID and column names of markers.}

\item{sep}{allele separator in the imported genotype data. Note: when using the special character like "|", remember to protect it as "\\|"(default).}
}
\value{
a dataframe of heterozygosity.0 is homozygous;1 is heterozygous. Each row denotes each individual; Each column denotes each locus.
}
\description{
Test heterozygosity at each locus
}
\details{
This function test the heterozygosity of each individuals at each locus.Output a table and Usually followed by write.csv(as.data.frame(y),file = "~/*.csv") to export the results.
}
\examples{
x <- data.frame(STR1=c("12|12","13|14","13|13","14|15"),
               SNP1=c("A|A","T|T","A|T","A|T"))
Heterozygous(x,"\\\\|")

}
